requirements.txt
asgiref==3.2.10
autopep8==1.5.5
boto3==1.15.14
botocore==1.18.14
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.25.10
whitenoise==5.2.0
